# `lint` only finds targets and files if the relevant linters were specified (Cherry-pick of #15579)